+++ src/main/java/org/codehaus/mojo/versions/UseLatestReleasesMojo.java
@@ -26,18 +26,15 @@ import org.apache.maven.artifact.metadata.ArtifactMetadataRetrievalException;
 import org.apache.maven.artifact.versioning.ArtifactVersion;
 import org.apache.maven.artifact.versioning.DefaultArtifactVersion;
 import org.apache.maven.artifact.versioning.VersionRange;
-import org.apache.maven.model.Dependency;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.codehaus.mojo.versions.api.ArtifactVersions;
+import org.codehaus.mojo.versions.change.VersionChange;
 import org.codehaus.mojo.versions.ordering.MajorMinorIncrementalFilter;
-import org.codehaus.mojo.versions.rewriting.ModifiedPomXMLEventReader;
 
-import javax.xml.stream.XMLStreamException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -86,28 +83,19 @@ public class UseLatestReleasesMojo extends ParentUpdatingDependencyUpdateMojo
     // ------------------------------ METHODS --------------------------
 
     @Override
-    protected void setVersions(ModifiedPomXMLEventReader pom, Collection<Dependency> dependencies)
-          throws ArtifactMetadataRetrievalException, XMLStreamException, MojoExecutionException
+    Collection<VersionChange> getVersionChanges(Collection<ArtifactIdentifier> artifacts) throws MojoExecutionException, ArtifactMetadataRetrievalException
     {
-        useLatestReleases(pom, dependencies);
-    }
+        final Collection<VersionChange> versionsToChange = new ArrayList<>();
 
-    private void useLatestReleases( ModifiedPomXMLEventReader pom, Collection<Dependency> dependencies )
-        throws XMLStreamException, MojoExecutionException, ArtifactMetadataRetrievalException
-    {
         int segment = determineUnchangedSegment( allowMajorUpdates, allowMinorUpdates, allowIncrementalUpdates );
         MajorMinorIncrementalFilter majorMinorIncfilter =
             new MajorMinorIncrementalFilter( allowMajorUpdates, allowMinorUpdates, allowIncrementalUpdates );
 
-        Iterator<Dependency> i = dependencies.iterator();
-
-        while ( i.hasNext() )
+        for ( ArtifactIdentifier dep : artifacts )
         {
-            Dependency dep = i.next();
-
             if ( isExcludeReactor() && isProducedByReactor( dep ) )
             {
-                getLog().info( "Ignoring reactor dependency: " + toString( dep ) );
+                getLog().info( "Ignoring reactor dependency: " + dep );
                 continue;
             }
 
@@ -115,7 +103,7 @@ public class UseLatestReleasesMojo extends ParentUpdatingDependencyUpdateMojo
             Matcher versionMatcher = matchSnapshotRegex.matcher( version );
             if ( !versionMatcher.matches() )
             {
-                Artifact artifact = this.toArtifact( dep );
+                Artifact artifact = dep.getArtifact( getProject(), getHelper() );
                 if ( !isIncluded( artifact ) )
                 {
                     continue;
@@ -124,7 +112,7 @@ public class UseLatestReleasesMojo extends ParentUpdatingDependencyUpdateMojo
                 ArtifactVersion selectedVersion = new DefaultArtifactVersion( version );
                 getLog().debug( "Selected version:" + selectedVersion.toString() );
 
-                getLog().debug( "Looking for newer versions of " + toString( dep ) );
+                getLog().debug( "Looking for newer versions of " + dep );
                 ArtifactVersions versions = getHelper().lookupArtifactVersions( artifact, false );
                 ArtifactVersion[] newer = versions.getNewerVersions( version, segment, false );
                 newer = filterVersionsWithIncludes( newer, artifact );
@@ -132,10 +120,12 @@ public class UseLatestReleasesMojo extends ParentUpdatingDependencyUpdateMojo
                 ArtifactVersion[] filteredVersions = majorMinorIncfilter.filter( selectedVersion, newer );
                 if ( filteredVersions.length > 0 )
                 {
-                    setVersion(pom, dep, version, artifact, filteredVersions[filteredVersions.length - 1]);
+                    final VersionChange versionChange = new VersionChange( artifact.getGroupId(), artifact.getArtifactId(), version, filteredVersions[filteredVersions.length - 1].toString() );
+                    versionsToChange.add( versionChange );
                 }
             }
         }
+        return versionsToChange;
     }
 
     private ArtifactVersion[] filterVersionsWithIncludes( ArtifactVersion[] newer, Artifact artifact )