+++ src/main/java/org/codehaus/mojo/versions/UseLatestSnapshotsMojo.java
@@ -23,20 +23,17 @@ import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.metadata.ArtifactMetadataRetrievalException;
 import org.apache.maven.artifact.versioning.ArtifactVersion;
 import org.apache.maven.artifact.versioning.DefaultArtifactVersion;
-import org.apache.maven.model.Dependency;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.codehaus.mojo.versions.api.ArtifactVersions;
+import org.codehaus.mojo.versions.change.VersionChange;
 import org.codehaus.mojo.versions.ordering.MajorMinorIncrementalFilter;
 import org.codehaus.mojo.versions.ordering.VersionComparator;
-import org.codehaus.mojo.versions.rewriting.ModifiedPomXMLEventReader;
 
-import javax.xml.stream.XMLStreamException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Iterator;
-import java.util.ArrayList;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -85,28 +82,19 @@ public class UseLatestSnapshotsMojo
     // ------------------------------ METHODS --------------------------
 
     @Override
-    protected void setVersions(ModifiedPomXMLEventReader pom, Collection<Dependency> dependencies)
-          throws ArtifactMetadataRetrievalException, XMLStreamException, MojoExecutionException
+    Collection<VersionChange> getVersionChanges(Collection<ArtifactIdentifier> artifacts) throws MojoExecutionException, ArtifactMetadataRetrievalException
     {
-        useLatestSnapshots(pom, dependencies);
-    }
+        final Collection<VersionChange> versionsToChange = new ArrayList<>();
 
-    private void useLatestSnapshots( ModifiedPomXMLEventReader pom, Collection<Dependency> dependencies )
-        throws XMLStreamException, MojoExecutionException, ArtifactMetadataRetrievalException
-    {
         int segment = determineUnchangedSegment( allowMajorUpdates, allowMinorUpdates, allowIncrementalUpdates );
         MajorMinorIncrementalFilter majorMinorIncfilter =
                 new MajorMinorIncrementalFilter( allowMajorUpdates, allowMinorUpdates, allowIncrementalUpdates );
 
-        Iterator<Dependency> i = dependencies.iterator();
-
-        while ( i.hasNext() )
+        for ( ArtifactIdentifier dep : artifacts )
         {
-            Dependency dep = i.next();
-
             if ( isExcludeReactor() && isProducedByReactor( dep ) )
             {
-                getLog().info( "Ignoring reactor dependency: " + toString( dep ) );
+                getLog().info( "Ignoring reactor dependency: " + dep );
                 continue;
             }
 
@@ -114,8 +102,8 @@ public class UseLatestSnapshotsMojo
             Matcher versionMatcher = matchSnapshotRegex.matcher( version );
             if ( !versionMatcher.matches() )
             {
-                getLog().debug( "Looking for latest snapshot of " + toString( dep ) );
-                Artifact artifact = this.toArtifact( dep );
+                getLog().debug( "Looking for latest snapshot of " + dep );
+                Artifact artifact = dep.getArtifact( getProject(), getHelper() );
                 if ( !isIncluded( artifact ) )
                 {
                     continue;
@@ -128,12 +116,12 @@ public class UseLatestSnapshotsMojo
                 final DefaultArtifactVersion lowerBound = new DefaultArtifactVersion( version );
                 if ( segment + 1 > versionComparator.getSegmentCount( lowerBound ) )
                 {
-                    getLog().info( "Ignoring " + toString( dep ) + " as the version number is too short" );
+                    getLog().info( "Ignoring " + dep + " as the version number is too short" );
                     continue;
                 }
                 ArtifactVersion upperBound =
                     segment >= 0 ? versionComparator.incrementSegment( lowerBound, segment ) : null;
-                getLog().info( "Upper bound: " + ( upperBound == null ? "none" : upperBound.toString() ) );
+                getLog().info( "Upper bound: " + (upperBound == null ? "none" : upperBound.toString()) );
                 ArtifactVersion[] newer = versions.getVersions( lowerBound, upperBound, true, false, false );
                 getLog().debug( "Candidate versions " + Arrays.asList( newer ) );
 
@@ -144,21 +132,23 @@ public class UseLatestSnapshotsMojo
                     String newVersion = newer[j].toString();
                     if ( matchSnapshotRegex.matcher( newVersion ).matches() )
                     {
-                        snapshotsOnly.add(newer[j]);
+                        snapshotsOnly.add( newer[j] );
                     }
                 }
-                getLog().debug( "Snapshot Only versions " + snapshotsOnly.toString());
+                getLog().debug( "Snapshot Only versions " + snapshotsOnly.toString() );
 
-                ArtifactVersion[] filteredVersions = majorMinorIncfilter.filter( selectedVersion,(ArtifactVersion[]) snapshotsOnly.toArray(new ArtifactVersion[snapshotsOnly.size()]));
+                ArtifactVersion[] filteredVersions = majorMinorIncfilter.filter( selectedVersion, (ArtifactVersion[]) snapshotsOnly.toArray( new ArtifactVersion[snapshotsOnly.size()] ) );
                 getLog().debug( "Filtered versions " + Arrays.asList( filteredVersions ) );
 
 
                 if ( filteredVersions.length > 0 )
                 {
-                    setVersion(pom, dep, version, artifact, filteredVersions[filteredVersions.length - 1]);
+                    final VersionChange versionChange = new VersionChange( artifact.getGroupId(), artifact.getArtifactId(), version, filteredVersions[filteredVersions.length - 1].toString() );
+                    versionsToChange.add( versionChange );
                 }
             }
         }
+        return versionsToChange;
     }
 
 }